Fixed ComBinlogDump pos field overflow
The pos field in ComBinlogDump is a 32 bit unsigned integer, but a real
binlog position can be larger than 4GB, which will cause the pos field
to overflow. This commit fixes the overflow by using a 64 bit unsigned.

+++ src/packets/binlog_request.rs
@@ -134,7 +134,7 @@ impl<'a> BinlogRequest<'a> {
             Either::Right(cmd)
         } else {
             let cmd = ComBinlogDump::new(self.server_id)
-                .with_pos(self.pos as u32)
+                .with_pos(self.pos)
                 .with_filename(&*self.filename)
                 .with_flags(self.flags & BinlogDumpFlags::BINLOG_DUMP_NON_BLOCK);
             Either::Left(cmd)
